feat(derive): seperate xcm && outcome errors
