+++ primitives/derive/src/xcm/mod.rs
@@ -1,4 +1,6 @@
 //! XCM macros
+mod outcome;
 mod result;
+mod xcm;
 
 pub use result::error;
